Reorganize tests
